Move theme selector to the correct spot
